% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_box.R
\name{geo_bbox}
\alias{geo_bbox}
\title{Bounding box}
\usage{
geo_bbox(x)
}
\arguments{
\item{x}{input}
}
\description{
Bounding box
}
\examples{
library("lawn")
library("geojson")
geo_bbox(as.geojson(lawn_data$polygons_count))

# x <- as.geojson(lawn_data$polygons_count)
# geojson::geo_bbox(x)
}

