% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{geo_circle}
\alias{geo_circle}
\title{Takes a \link{Point} and calculates the circle polygon given a radius
in degrees, radians, miles, or kilometers; and steps for precision.}
\usage{
geo_circle(center, radius, steps, units = "kilometers")
}
\arguments{
\item{center}{\link{Feature}<\link{Point}> center point}

\item{radius}{(numeric) radius of the circle}

\item{steps}{(numeric) number of steps. default: 64}

\item{units}{(character) miles, kilometers, degrees, or
radians}
}
\value{
\link{Feature}<\link{Polygon}> circle polygon
}
\description{
Takes a \link{Point} and calculates the circle polygon given a radius
in degrees, radians, miles, or kilometers; and steps for precision.
}
\examples{
# circle(center, radius, steps, units)
}
