% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{geo_distance}
\alias{geo_distance}
\title{Calculate distance between two GeoJSON points}
\usage{
geo_distance(from, to, units)
}
\arguments{
\item{from}{Origin point}

\item{to}{Destination point}

\item{units}{(character) Can be degrees, radians, miles, or kilometers}
}
\value{
Single numeric value
}
\description{
Calculate distance between two GeoJSON points
}
\examples{
point1 <- '{
  "type": "Feature",
  "properties": {
    "marker-color": "#f00"
   },
   "geometry": {
      "type": "Point",
      "coordinates": [-75.343, 39.984]
   }
}'

point2 <- '{
  "type": "Feature",
  "properties": {
     "marker-color": "#0f0"
   },
   "geometry": {
      "type": "Point",
      "coordinates": [-75.534, 39.123]
    }
}'

geo_distance(point1, point2, units = "miles")
geo_distance(point1, point2, units = "kilometers")
geo_distance(point1, point2, units = "radians")

pt1 <- '{
  "type": "Feature",
  "geometry": {
    "type": "Point",
    "coordinates": [
       0.5,
       0.5
     ]
   }
}'

pt2 <- '{
  "type": "Feature",
  "geometry": {
    "type": "Point",
    "coordinates": [2, 2]
   }
}'

geo_distance(pt1, pt2, units = "miles")

# comparison to rgeos
# the geojson points and WKT points are the same, just diff. formats
library(rgeos)
library(microbenchmark)
rgeospt1 <- readWKT("POINT(0.5 0.5)")
rgeospt2 <- readWKT("POINT(2 2)")
microbenchmark::microbenchmark(
  rgeos = rgeos::gDistance(rgeospt1, rgeospt2),
  geoops = geoops::geo_distance(pt1, pt2, units = "miles"),
  times = 1000L
)
}

