% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{geo_get_coords}
\alias{geo_get_coords}
\title{Get coordinates}
\usage{
geo_get_coords(x)
}
\arguments{
\item{x}{geojson string}
}
\value{
a numeric vector with coordinates
}
\description{
Get coordinates
}
\examples{
\dontrun{
x <- '{"type": "Feature", "geometry": {"type": "Point","coordinates": [1, 2]},"properties": {}}'
geo_get_coords(x)
x <- '{"type": "Point", "coordinates": [1, 2]}'
geo_get_coords(x)
x <- '[0, 5]'
geo_get_coords(x)
# x <- '[5]'
# geo_get_coords(x)
}
}
