% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglegrid.R
\name{geo_trianglegrid}
\alias{geo_trianglegrid}
\title{Trianglegrid}
\usage{
geo_trianglegrid(bbox, cell_size, units = "kilometers")
}
\arguments{
\item{bbox}{extent in \verb{[minX, minY, maxX, maxY]} order}

\item{cell_size}{(numeric) the distance across each cell}

\item{units}{(character) used in calculating cellSize, can be degrees,
radians, miles, or kilometers (default)}
}
\value{
\link{FeatureCollection}<\link{Polygon}> grid of polygons
}
\description{
Takes a bounding box and a cell depth and returns a set of \link{Polygon}'s
in a grid.
}
\examples{
geo_trianglegrid(c(-77.3876, 38.7198, -76.9482, 39.0277), 3, "miles")
geo_trianglegrid(c(-77.3876, 38.7198, -76.9482, 39.0277), 10, "miles")
geo_trianglegrid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, "miles")
}
