% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson-types.R
\name{geojson-types}
\alias{geojson-types}
\alias{data-Point}
\alias{data-MultiPoint}
\alias{data-Polygon}
\alias{data-MultiPolygon}
\alias{data-LineString}
\alias{data-MultiLineString}
\alias{data-Feature}
\alias{data-FeatureCollection}
\alias{data-GeometryCollection}
\alias{data-GeoJSON}
\title{Description of GeoJSON data types}
\description{
Description of GeoJSON data types
}
\section{GeoJSON object}{


GeoJSON always consists of a single object. This object (referred to as the
GeoJSON object below) represents a geometry, feature, or collection of
features.

\itemize{
\item The GeoJSON object may have any number of members (name/value pairs).
\item The GeoJSON object must have a member with the name "type". This
member's value is a string that determines the type of the GeoJSON object.
\item The value of the type member must be one of: "Point", "MultiPoint",
"LineString", "MultiLineString", "Polygon", "MultiPolygon",
"GeometryCollection", "Feature", or "FeatureCollection". The case of the
type member values must be as shown here.
\item A GeoJSON object may have an optional "crs" member, the value of
which must be a coordinate reference system object (see 3. Coordinate
Reference System Objects).
\item A GeoJSON object may have a "bbox" member, the value of which must
be a bounding box array (see 4. Bounding Boxes).
}
}

\section{Point}{


For type "Point", the "coordinates" member must be a single position.

Example JSON:\preformatted{{
"type": "Point",
"coordinates": [100.0, 0.0]
}
}
}

\section{MultiPoint}{


For type "MultiPoint", the "coordinates" member must be an array of
positions.

Example JSON:\preformatted{{
"type": "MultiPoint",
"coordinates": [
   [100.0, 0.0],
   [101.0, 1.0]
 ]
}
}

[
[100.0, 0.0]: R:%0A%20%20%20[100.0,%200.0
[101.0, 1.0]: R:101.0,%201.0
}

\section{Polygon}{


For type "Polygon", the "coordinates" member must be an array of LinearRing
coordinate arrays. For Polygons with multiple rings, the first must be the
exterior ring and any others must be interior rings or holes.

Example JSON:\preformatted{{
"type": "Polygon",
"coordinates": [
  [
    [100.0, 0.0],
    [101.0, 0.0],
    [101.0, 1.0],
    [100.0, 1.0],
    [100.0, 0.0]
   ]
 ]
}
}

[
[
[100.0, 0.0]: R:%0A%20%20[%0A%20%20%20%20[100.0,%200.0
[101.0, 0.0]: R:101.0,%200.0
[101.0, 1.0]: R:101.0,%201.0
[100.0, 1.0]: R:100.0,%201.0
[100.0, 0.0]: R:100.0,%200.0
}

\section{MultiPolygon}{


For type "MultiPolygon", the "coordinates" member must be an array of
Polygon coordinate arrays.

Example JSON:\preformatted{{
"type": "MultiPolygon",
"coordinates": [
  [[[102.0, 2.0], [103.0, 2.0], [103.0, 3.0], [102.0, 3.0], [102.0, 2.0]]],
  [[[100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0]],
  [[100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2]]]
  ]
}
}

[
[[[102.0, 2.0]: R:%0A%20%20[[[102.0,%202.0
[103.0, 2.0]: R:103.0,%202.0
[103.0, 3.0]: R:103.0,%203.0
[102.0, 3.0]: R:102.0,%203.0
[102.0, 2.0]: R:102.0,%202.0
[[[100.0, 0.0]: R:[[100.0,%200.0
[101.0, 0.0]: R:101.0,%200.0
[101.0, 1.0]: R:101.0,%201.0
[100.0, 1.0]: R:100.0,%201.0
[100.0, 0.0]: R:100.0,%200.0
[[100.2, 0.2]: R:[100.2,%200.2
[100.8, 0.2]: R:100.8,%200.2
[100.8, 0.8]: R:100.8,%200.8
[100.2, 0.8]: R:100.2,%200.8
[100.2, 0.2]: R:100.2,%200.2
}

\section{LineString}{


For type "LineString", the "coordinates" member must be an array of two or
more positions. A LinearRing is closed LineString with 4 or more positions.
The first and last positions are equivalent (they represent equivalent
points). Though a LinearRing is not explicitly represented as a GeoJSON
geometry type, it is referred to in the Polygon geometry type definition.

Example JSON:\preformatted{{
"type": "LineString",
"coordinates": [
   [100.0, 0.0],
   [101.0, 1.0]
 ]
}
}

[
[100.0, 0.0]: R:%0A%20%20%20[100.0,%200.0
[101.0, 1.0]: R:101.0,%201.0
}

\section{MultiLineString}{


For type "MultiLineString", the "coordinates" member must be an array of
LineString coordinate arrays.

Example JSON:\preformatted{{
 "type": "MultiLineString",
 "coordinates": [
     [[ -105, 39 ], [ -105, 39 ]],
     [[ -105, 39 ], [ -105, 39 ]]
  ]
}
}

[
[[ -105, 39 ]: R:%0A%20%20%20%20%20[[%20-105,%2039%20
[ -105, 39 ]: R:%20-105,%2039%20
[[ -105, 39 ]: R:[%20-105,%2039%20
[ -105, 39 ]: R:%20-105,%2039%20
}

\section{Feature}{


A GeoJSON object with the type "Feature" is a feature object:
\itemize{
\item A feature object must have a member with the name "geometry". The
value of the geometry member is a geometry object as defined above or a
JSON null value.
\item A feature object must have a member with the name "properties". The
value of the properties member is an object (any JSON object or a JSON null
value).
\item If a feature has a commonly used identifier, that identifier should
be included as a member of the feature object with the name "id".
}

Example JSON:\preformatted{{
  "type": "Feature",
  "properties": {
    "population": 200
  },
  "geometry": {
    "type": "Point",
    "coordinates": [10.724029, 59.926807]
  }
}
}
}

