% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_play.R, R/get_coord.R
\name{get_coord}
\alias{get_coord}
\alias{get_json}
\alias{get_str}
\alias{in_ring}
\alias{parse_bools}
\alias{parse_ifelse}
\alias{parse_type}
\title{Get coordinates}
\usage{
get_coord(x)

get_str(x)

get_json()

parse_type(x)

parse_ifelse(x)

parse_bools(x)

in_ring(pt, ring)

get_coord(x)
}
\arguments{
\item{x}{stuff}

\item{x}{geojson string}
}
\description{
Get coordinates

Get coordinates
}
\examples{
\dontrun{
x <- '{"type": "Feature", "geometry": {"type": "Point","coordinates": [1, 2]},"properties": {}}'
get_coord(x)
x <- '{"type": "Point", "coordinates": [1, 2]}'
get_coord(x)
x <- '[0, 5]'
get_coord(x)

# get_coord(x)
# get_str(x)
# get_json()
# parse_type(x)

# in ring
pt <- "[-77, 44]"
ring <- "[[10,59],[10.1,59.8],[10.3,60],[10.3,59],[10,59]]"
in_ring(pt, ring)

pt <- "[10.17, 59.4]"
ring <- "[[10,59],[10.1,59.8],[10.3,60],[10.3,59],[10,59]]"
in_ring(pt, ring)
}
\dontrun{
x <- '{"type": "Feature", "geometry": {"type": "Point","coordinates": [1, 2]},"properties": {}}'
get_coord(x)
x <- '{"type": "Point", "coordinates": [1, 2]}'
get_coord(x)
x <- '[0, 5]'
get_coord(x)
}
}

