% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{get_coords}
\alias{get_coords}
\title{Get coordinates}
\usage{
get_coords(x)
}
\arguments{
\item{x}{geojson string}
}
\value{
a string, with coordinates as a JSON array
}
\description{
Get coordinates
}
\examples{
\dontrun{
x <- '{"type": "Feature", "geometry": {"type": "Point","coordinates": [1, 2]},"properties": {}}'
get_coords(x)
x <- '{"type": "Point", "coordinates": [1, 2]}'
get_coords(x)
x <- '[0, 5]'
get_coords(x)
}
}
