% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_ring.R
\name{in_ring}
\alias{in_ring}
\title{Get coordinates}
\usage{
in_ring(pt, ring)
}
\arguments{
\item{pt}{a point}

\item{ring}{a ring}
}
\description{
Get coordinates
}
\examples{
\dontrun{
pt <- "[-77, 44]"
ring <- "[[10,59],[10.1,59.8],[10.3,60],[10.3,59],[10,59]]"
in_ring(pt, ring)

pt <- "[10.17, 59.4]"
ring <- "[[10,59],[10.1,59.8],[10.3,60],[10.3,59],[10,59]]"
in_ring(pt, ring)
}
}

