% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineString.R
\name{LineString}
\alias{LineString}
\title{LineString}
\description{
For type "LineString", the "coordinates" member must be an array of two or
more positions. A LinearRing is closed LineString with 4 or more positions.
The first and last positions are equivalent (they represent equivalent
points). Though a LinearRing is not explicitly represented as a GeoJSON
geometry type, it is referred to in the Polygon geometry type definition.
}
\examples{
'{
   "type": "LineString",
   "coordinates": [
      [100.0, 0.0],
      [101.0, 1.0]
   ]
}'
}
\seealso{
Other geo types: 
\code{\link{FeatureCollection}},
\code{\link{Feature}},
\code{\link{GeometryCollection}},
\code{\link{MultiLineString}},
\code{\link{MultiPoint}},
\code{\link{MultiPolygon}},
\code{\link{Point}},
\code{\link{Polygon}},
\code{\link{geojson-types}}
}
\concept{geo types}
