% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLineString.R
\name{MultiLineString}
\alias{MultiLineString}
\title{MultiLineString}
\description{
For type "MultiLineString", the "coordinates" member must be an array of
LineString coordinate arrays.
}
\examples{
'{
  "type": "MultiLineString",
  "coordinates": [
     [[ -105, 39 ], [ -105, 39 ]],
     [[ -105, 39 ], [ -105, 39 ]]
  ]
}'
}
\seealso{
Other geo types: 
\code{\link{FeatureCollection}},
\code{\link{Feature}},
\code{\link{GeometryCollection}},
\code{\link{LineString}},
\code{\link{MultiPoint}},
\code{\link{MultiPolygon}},
\code{\link{Point}},
\code{\link{Polygon}},
\code{\link{geojson-types}}
}
\concept{geo types}
