% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPoint.R
\name{MultiPoint}
\alias{MultiPoint}
\title{GeoJSON MultiPoint}
\description{
For type "MultiPoint", the "coordinates" member must be an array of
positions.
}
\examples{
'{
  "type": "MultiPoint",
  "coordinates": [
     [100.0, 0.0],
     [101.0, 1.0]
  ]
}'
}
\seealso{
Other geo types: 
\code{\link{FeatureCollection}},
\code{\link{Feature}},
\code{\link{GeometryCollection}},
\code{\link{LineString}},
\code{\link{MultiLineString}},
\code{\link{MultiPolygon}},
\code{\link{Point}},
\code{\link{Polygon}},
\code{\link{geojson-types}}
}
\concept{geo types}
