% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPolygon.R
\name{MultiPolygon}
\alias{MultiPolygon}
\title{MultiPolygon}
\description{
For type "MultiPolygon", the "coordinates" member must be an array of
Polygon coordinate arrays.
}
\examples{
'{
  "type": "MultiPolygon",
  "coordinates": [
    [[[102.0, 2.0], [103.0, 2.0], [103.0, 3.0], [102.0, 3.0], [102.0, 2.0]]],
    [[[100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0]],
    [[100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2]]]
  ]
}'
}
\seealso{
Other geo types: 
\code{\link{FeatureCollection}},
\code{\link{Feature}},
\code{\link{GeometryCollection}},
\code{\link{LineString}},
\code{\link{MultiLineString}},
\code{\link{MultiPoint}},
\code{\link{Point}},
\code{\link{Polygon}},
\code{\link{geojson-types}}
}
\concept{geo types}
