% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Overlay}
\usage{
overlay(points, polygons)
}
\arguments{
\item{points}{geojson points}

\item{polygons}{geojson polygons}
}
\description{
Clip a set of points to a polygon
}
\examples{
library("lawn")
lawn_data$points_within \%>\% view
overlay(
  points = lawn_data$points_within,
  polygons = lawn_data$polygons_within) \%>\% view


}

