% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.R
\name{Point}
\alias{Point}
\title{GeoJSON Point}
\description{
For type "Point", the "coordinates" member must be a single position.
}
\examples{
'{
   "type": "Point",
   "coordinates": [100.0, 0.0]
}'
}
\seealso{
Other geo types: 
\code{\link{FeatureCollection}},
\code{\link{Feature}},
\code{\link{GeometryCollection}},
\code{\link{LineString}},
\code{\link{MultiLineString}},
\code{\link{MultiPoint}},
\code{\link{MultiPolygon}},
\code{\link{Polygon}},
\code{\link{geojson-types}}
}
\concept{geo types}
