% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{Polygon}
\alias{Polygon}
\title{Polygon}
\description{
For type "Polygon", the "coordinates" member must be an array of LinearRing
coordinate arrays. For Polygons with multiple rings, the first must be the
exterior ring and any others must be interior rings or holes.
}
\examples{
'{
"type": "Polygon",
"coordinates": [
  [
    [100.0, 0.0],
    [101.0, 0.0],
    [101.0, 1.0],
    [100.0, 1.0],
    [100.0, 0.0]
   ]
 ]
}'
}
\seealso{
Other geo types: 
\code{\link{FeatureCollection}},
\code{\link{Feature}},
\code{\link{GeometryCollection}},
\code{\link{LineString}},
\code{\link{MultiLineString}},
\code{\link{MultiPoint}},
\code{\link{MultiPolygon}},
\code{\link{Point}},
\code{\link{geojson-types}}
}
\concept{geo types}
