% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Get json library version information}
\usage{
version()
}
\value{
JSON as a string for now, will make a list soon
}
\description{
Get json library version information
}
\examples{
version()
}
