% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_box.R
\name{geo_bbox}
\alias{geo_bbox}
\title{Bounding box}
\usage{
geo_bbox(x)
}
\arguments{
\item{x}{input}
}
\description{
Bounding box
}
\examples{
x <- '{
   "type": "FeatureCollection",
   "features": [
       {
           "type": "Feature",
           "properties": {

           },
           "geometry": {
               "type": "Polygon",
               "coordinates": [
                   [
                       [
                           -112.072391,
                           46.586591
                       ],
                       [
                           -112.072391,
                           46.61761
                       ],
                       [
                           -112.028102,
                           46.61761
                       ],
                       [
                           -112.028102,
                           46.586591
                       ],
                       [
                           -112.072391,
                           46.586591
                       ]
                   ]
               ]
           }
       },
       {
           "type": "Feature",
           "properties": {

           },
           "geometry": {
               "type": "Polygon",
               "coordinates": [
                   [
                       [
                           -112.023983,
                           46.570426
                       ],
                       [
                           -112.023983,
                           46.615016
                       ],
                       [
                           -111.966133,
                           46.615016
                       ],
                       [
                           -111.966133,
                           46.570426
                       ],
                       [
                           -112.023983,
                           46.570426
                       ]
                   ]
               ]
           }
       }
   ]
}'
geo_bbox(x)
}
