% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{get_coord}
\alias{get_coord}
\title{Get coordinates}
\usage{
get_coord(x)
}
\arguments{
\item{x}{geojson string}
}
\description{
Get coordinates
}
\examples{
\dontrun{
x <- '{"type": "Feature", "geometry": {"type": "Point","coordinates": [1, 2]},"properties": {}}'
get_coord(x)
x <- '{"type": "Point", "coordinates": [1, 2]}'
get_coord(x)
x <- '[0, 5]'
get_coord(x)
}
}
